Update index.js
A middleware function has been implemented for the /customer/auth/* route. This function is responsible for handling authentication for routes under "/customer/auth/*" (Line 13).

The function "auth" takes three parameters: req (the request object), res (the response object), and next (a callback function to pass control to the next middleware function).

First, the function checks if there is an "authorization" property in the session object of the request (req.session.authorization). If it exists, it means that the user has previously logged in, and an access token should be stored in req.session.authorization['accessToken'] (Line 16). This token is then extracted and stored in the "token" variable.

The jwt.verify function is then invoked with the token and a secret key ("access"). This function verifies the token: if the token is valid, it decodes the payload of the token and passes it to the callback function as the "user" parameter (Line 17). The "user" object is then attached to the request object (req.user = user) (Line 19), and the next() function is called to pass control to the next middleware function or route handler.

If the token is not valid (for example, if it has been tampered with, or if it has expired), jwt.verify will pass an error to the callback function. In this case, the function responds with a 401 status code and a JSON object with a message stating "User not authenticated" (Line 23).

If the "authorization" property does not exist in the session object, it means that the user has not logged in. In this case, the function responds with a 403 status code and a JSON object with a message stating "User not logged in" (Line 27).