Add GET "book list" request in general.js
The new code from lines 14 to 18 defines a route handler for HTTP GET requests to the root URL ("/") using Express.

The line `public_users.get('/', function (req, res) {…});` sets up the route (lines 14 to 18). `public_users` is an instance of an Express router from `general.js`. The `.get` method is used to handle HTTP GET requests. The first argument to this method is the path for which the handler is invoked. In this case, it's the root path ('/').

The second argument of the `.get` method is a callback function that is invoked when a GET request is made to the specified path. This function takes two arguments: `req` and `res`. `req` is an object containing information about the incoming request, and `res` is an object for sending the HTTP response (Line 14).

Inside the callback function, `res.send(JSON.stringify(books, null, 4));` is called (line 17).

`res.send` is a method provided by Express to send a response back to the client. The argument to `res.send` is the data to be sent as the response.

`JSON.stringify(books,null,4)` converts the `books` object into a JSON string.

The `null` and `4` arguments are optional parameters that control how stringification is done. The `null` argument means that all properties of the "books" object will be included in the JSON string. The `4` argument means that the JSON string will be formatted with four spaces of indentation for readability (Line 17).

In short, the function of this code is to set a route that responds to HTTP GET requests at the root URL with a JSON string representing the "books" object.