Update general.js ( Add "GET Book details based on ISBN")
The code snippet `public_users.get('/isbn/:isbn', function (req, res) {…});` is establishing a route on Line 21. `public_users` is an instance of an Express router. 

The `.get` method is utilized to manage HTTP GET requests, with the first argument representing the invoked handler path, in this case, 'isbn/:isbn'. 

The ':isbn' segment of the path functions as a router parameter for capturing dynamic parts of the URL (Line 22).

The second argument of the `.get` method is a callback function that is executed when a GET request is made to the specified path. This function takes two arguments: `req` and `res` (Line 21).

`req` is an object containing information about the incoming request, while `res` is utilized for sending the HTTP response.

Within the callback function, `const isbn = req.params.isbn;` is used to extract the ISBN from the request parameters. `req.params` is an object with properties mapped to the named route parameters, so `req.params.isbn` will contain the actual ISBN from the URL.

The `if (books[isbn]) {…} else {…}` statement on Line 24 is an if-else statement that checks if a book with the provided ISBN exists in the book's object.

If a book is found, it sends that book as the response using `res.send(books[isbn]);` (Line 25).

If no book is found, it sends a response containing the message "Book does not exist" using `res.send({message: "Book does not exist"});` (Line 26).