Update general.js (Modified  GET Book details based on Author)
I encountered an error when testing the "GET Book Details based on author" function. The 'booksArray' was not defined as the same array in "Get book details based on ISBN".

To solve this, I updated the code by adding 'const booksArray = Object.values(books);' in line 43 to ensure that I am using the same array to access the author's property and string value.

I tested the code again and the function works properly.

I also included a 'decodeURIComponent' function to decode the author's name from the route parameters in line 39. This change will allow the inclusion of spaces and other special characters in the author's name in the URL.