Update general.js (Modified "GET book review by ISBN")
I found an issue when I was transferring the code into the GitHub repository. I accidentally typed books.isbn instead of  'book.isbn' so it has the "if (book) " function thinking I am referring to a different book property. Also, I  forgot to add the s in "book.reviews" in line 86. There is also an extra "=" in the 82nd line.

After fixing the grammar errors, I added "books.review.length > 0" so if a book is found, the code will check if the reviews property of the book is present or an empty array (Line 85).

If the reviews property is an empty array, the code sends a response with the message "Please add a review" (Line 88). If no book is found, it sends a response with the message "Book does not exist"( line 91).