Update general.js (Add public_users.post request)
I added a route handler for a POST request to the "/register" endpoint. This code from line 7 to line 36  handles registering public users to the bookshop.

When a POST request is made to this endpoint, the function is called with two arguments: 'req' (the request object) and 'res'  (the response object) (line 8). The request object contains information about the client's request, which in this case is the username and password.

The request body extracts the username and password by using the 'req.body.username' and 'req.body.password' respectively (lines 11 and 12).

Next, the code checks if both the 'username' and 'password' are provided (line 14). If they are, it calls the 'doesExist' function with the username as the argument (line 23). The  'doesExist' function checks if a user with the given username already exists with the 'users' array. This is done by filtering the users' array to create a new one to hold the users whose username matches the given username value (line 24). If the new array has more than 0 elements, it means a user with the given username exists, so the function returns true (line 28). If the new array is empty, it means no user with the given username exists, resulting in the function returning 'false' (line 30).

If the 'doesExist' function returns false, it means no user with the given username exists. In that case, the code adds a new user with the username and password to the users' array. After that, it sends a response with the status code 200 and a message saying "User successfully registered. Now, you can log in" (line 17).

If the 'doesExist' function returns true, then a user with the given username already exists, so the code sends a response with the status code 404 and a message saying "User already exists!" (line 19).

If either the username or password is not provided, the code sends a response with the status code 404 and a message saying "Unable to register user" (line 35).

To summarize, the code from lines 7 to 36, handles the registration of public users to the bookshop when a POST request is made to the "/register" endpoint. It checks if the username and password are provided, then verifies if the username already exists. If the username is new, it adds the user to the database; otherwise, it returns an error message. If either the username or password is missing, it sends a relevant error message.