Update general.js (Fixed typos in the code)
PUBLIC_USERS.POST REQUEST ERROR:

After I run the 'public_users.post request'  I found an error where I get a message saying "Cannot access 'doesExist' before initialization". I looked into it and discovered I was trying to use the 'doesExist' function before it was defined.

In the previous commit push, I would call 'doesExist(username)' in the 'if' statement before the 'doesExist' function is defined in line 23. 

When I tried to execute 'doesExist(username)',   the code did not know what 'doesExist' was because the definition had not been reached yet. This is because JavaScript is a single-pass language. 

So to solve those issues, I had to move the 'doesExist' function from the space between lines 24 and 33  to the space between lines 13 and 22 instead. That way, the 'doesExist' function is defined before it is used, removing any chances of getting the "Cannot access 'doesExist' before initialization error'

CORRECT TYPOS:

While solving the problem with the 'doesExist' function, I found two typos in other parts of my code. These typos caused two important parts of the code to be undefined, affecting the code's function.

The first typo was in the "if doesExist" function where the name was misspelled as "doesExists". The second typo was in the definition for the function to find a book review based on the ISBN. The 'book.isbn' definition was mistyped as 'books.isbn' in the 105th line.

After I fixed those typos, the code began functioning properly again.