Update general.js (Add additional error messages)
I have also added extra error messages to the post registers_users function for specific scenarios, such as when a username or password is not provided. In these cases, the function will send a user response with the status code 400 and an appropriate error message.

In lines 24 and 25, the function sends a user response indicating that both a username and a password are required before the user can register.

In lines 26 and 27, the user response indicates that the password has been added, but a username is still required.

In lines 28 and 29, the username is added, but a message is sent saying the password is still required.

In lines 30, 31, and 32  the code uses the 'doesExist' function to check if the username already exists. If the username does exist, the username and password are pushed in, and a message is returned that says "User successfully registered. Now you can log in" in line 32.

In lines 33 and 34, the code checks if the registered user already exists before responding with a message saying "User already exists."

In lines 35 and 36, if the code is unable to register the user, it sends a response with the message "Unable to register user."

In a later post, I will add a feature that will check if the username and password are valid using the isValid function.