Update auth_users.js (Added registered users login)
In this updated code I am using the imported "jsonwebtoken" module, to create JWTs (line 2).  In the '/login' route in line 17,  the code checks if a user with the provided username and password exists in the users' array (lines 18, 19, and 21). 

If a user exists, a JWT is created with the username as the payload and "access" as the secret key (lines 23, 24, and 25). The created JWT is then saved in the session and is sent in the response (line 26).

If the user does not exist, a response with the status 401 and the message "Invalid username or password" is sent (line 28). 

Finally, the app.use function mounts the 'reged_users' router at the /customer/login path. This means the '/login' route will be accessible in the '/customer/login' endpoint (line 33).