Update auth_users.js (Updated isValid function)
At first, the isValid function was working but  it turns out, I made a mistake by assuming that "alphanumeric" refers to the username having both letters and numbers. But in reality, alphanumeric is the classification for numbers and letters.  

With this knowledge, I decided to modify the function to have a more specific requirement. Now the username must be six characters long and have at least one number and one letter.

The new features '/[a-z]+/i.test(username) ' value checks if the username has at least one letter, and '/[0-9]+/.test(username)' checks if the username has at least one number.