Update general.js (Update Get "Book reviews based on ISBN")
After implementing the "adding or updating a review" code, I discovered that when I use the get request for the book reviews based on ISBN, the response will always be "Please add a review" even when I tested with a review placed in the book database.

So to fix it, I added the function 'Object.keys(book.reviews).length === 0' to check if the reviews array is empty. 

Now, I can use the get request to bring back any reviews currently stored in the reviews array.