Update auth_users.js (Add "Delete Review" function)
I added the function that enables registered users to delete their reviews in the bookdb.js. The logged-in username will be retrieved from the session, and the ISBN will be used to locate the book with the review.

The auth/review endpoint/:isbn endpoint ensures only the registered username can delete their review and not someone else's by accident.