Update general.js ( Created a promise for book details based on the author)
Like the isbn function, the code for getting the book details can now be retrieved based on a promise in the callback function. Inside, the code searches for a book with the matching author and if the book is found, the promise is resolved.

But if the book is not found, the promise is rejected with an error indicating that the author is not listed.